split programfactory into hpp/cpp + cleanup
