+++ ttnn/cpp/ttnn/operations/generic/device/generic_op_device_operation.hpp
@@ -10,54 +10,17 @@
 
 #include "ttnn/decorators.hpp"
 #include "ttnn/tensor/tensor.hpp"
-#include "ttnn/device_operation.hpp"
+#include "generic_op_program_factory.hpp"
+#include "generic_op_device_operation_types.hpp"
 
 namespace ttnn::operations::generic {
 
 struct GenericOpDeviceOperation {
-    using operation_attributes_t = tt::tt_metal::experimental::MeshProgramDescriptor;
-
-    using tensor_return_value_t = Tensor;
-
-    using spec_return_value_t = TensorSpec;
-
-    // NOTE: output tensor is the last element in the vector io_tensors
-    struct tensor_args_t {
-        const std::vector<Tensor>& io_tensors;
-        const Tensor& output_tensor;
-    };
-
-    struct GenericMeshProgram {
-        struct shared_variables_t {
-            uint32_t num_kernel_handles{};
-            std::vector<tt::tt_metal::CBHandle> cb_handles;
-        };
-
-        struct mesh_shared_variables_t {
-            shared_variables_t program_shared_variables;
-        };
-
-        using cached_mesh_workload_t = ttnn::device_operation::AdaptedCachedMeshWorkload<mesh_shared_variables_t>;
-
-        static cached_mesh_workload_t create_mesh_workload(
-            const operation_attributes_t& operation_attributes,
-            const ttnn::MeshCoordinateRangeSet& tensor_coords,
-            const tensor_args_t& tensor_args,
-            tensor_return_value_t& tensor_return_value);
-
-        static ttnn::device_operation::CachedProgram<shared_variables_t> create_at(
-            const tt::tt_metal::ProgramDescriptor& program_descriptor,
-            const tensor_args_t& tensor_args,
-            tensor_return_value_t& tensor_return_value);
-
-        static void override_runtime_arguments(
-            cached_mesh_workload_t& cached_mesh_workload,
-            const operation_attributes_t& operation_attributes,
-            const tensor_args_t& tensor_args,
-            tensor_return_value_t& tensor_return_value);
-    };
-
-    using program_factory_t = std::variant<GenericMeshProgram>;
+    using operation_attributes_t = generic::operation_attributes_t;
+    using tensor_args_t = generic::tensor_args_t;
+    using spec_return_value_t = generic::spec_return_value_t;
+    using tensor_return_value_t = generic::tensor_return_value_t;
+    using program_factory_t = std::variant<program::GenericMeshProgramFactory>;
 
     static spec_return_value_t compute_output_specs(const operation_attributes_t&, const tensor_args_t&);
     static tensor_return_value_t create_output_tensors(const operation_attributes_t&, const tensor_args_t&);
@@ -75,7 +38,7 @@ struct GenericOpDeviceOperation {
 }  // namespace ttnn::operations::generic
 
 namespace ttnn::prim {
-ttnn::operations::generic::GenericOpDeviceOperation::tensor_return_value_t generic_op(
+ttnn::operations::generic::tensor_return_value_t generic_op(
     const std::vector<Tensor>& io_tensors,
-    const ttnn::operations::generic::GenericOpDeviceOperation::operation_attributes_t& operation_attributes);
+    const ttnn::operations::generic::operation_attributes_t& operation_attributes);
 }  // namespace ttnn::prim
